feat!: enable msm gpu algorithm selection
`BellmanMSM` and `CUZK` are MSM gpu algorithms. I meant to replace
BellmanMSM with CUZK because it ran very slow when the scalar fields
are not randomly distributed. But it turns out that CUZK is much slower
even in a randomly distributed case at this moment. While figuring out
a reason, it supports both apis.

TO DO: Provide c apis for algorithm selection.

BREAKING CHANGE: static methods of VariableBaseMSMGpu are removed.
Instead, it has methods to run with a uniform interface.