feat!(math): do not remove high degree zeros for `UnivariateEvaluations`
If we remove high degree zeros, this results in confusion. For example,
if this is constructed with a 2-degree evaluations vector `[1, 2, 0]` and
this removes the last element, then it would be `[1, 2]`. This is
1-degree polynomial! To prevent it from happening, the evaluations
should be remained untouched.

BREAKING CHANGE: This changes semantics of `UnivariateEvaluations`,
the result of `Zero()`, `One()`, `IsZero()`, `IsOne()` and `Degree()`
methods could be different.