refac(zk): move transition func from `Argument` to `Permuted` for lookup
This commit gathers the transition functions of the lookup
into the `LookupRunner` to enhance the readability of the protocol.

And the `PermutedArgument()` has been modified to consume the input `Argument`.
This is because there is no need to keep the already passed value (`Argument`)
of the running state in memory.