feat: generalize transcript
Current `Transcript` assumes that its commitment type is an affine
point. This commit breaks this assumption and it differentiates cases.

1. Commitment and Scalar are same types, such as FRI commitment cases.
2. Commitment and Scalar are different types, such as KZG commitment cases.

Also `TranscriptReader` and `TranscriptWriter` child classes uses
`Curve` type as a template parameter which is different from
`Commitment` type from a parent class. This commit unifies template
parameter for consistency as well.

Other than above, it does as follows.

- move halo2 specific things to `zk/plonk/halo2`.
- remove `Challenge255` which leads to removing
  `SqueezeChallengeAsScalar()`.
- rename `ReadScalar()` and `ReadPoint()` to `ReadFromProof()`
  for consistency with `WriteToProof()` and `WriteToTranscript()`.
- remove unused headers from transcript implementation.