build(math): apply string flag when generating prime fields
This is needed because we found that subgroup generator of Halo2 bn254
scalar field is different from ours. But we don't want to change our
default subgroup generator just because we want to make it produce the
same result with Halo2's. So this commit allows us to change not only
subgroup generator but also other properties that might be different
from each implementation easily.

`xx_small_group_adicity`, `xx_small_group_base_` and `xx_subgroup_generator`
are added.