feat(c/zk): enable logging transcript state to restore transcript state
Unlike the RNG state, guessing the value of this component is impossible.
This is because the instance values must be transcribed, and all we have
are univariate evaluations with a size equal to the domain size.
Determining whether the zero value represents a genuine instance
or an empty one is challenging. To address this, additional data
must be logged for the transcript state.

A new environment variable has been introduced.

- `TACHYON_TRANSCRIPT_STATE_LOG_PATH`: When activated, the serialized
  state of the Transcript is saved to the speicified path.