fix(math): correct parallel lagrange interpolation
When openmp is activated, two bugs are present in the Lagrange interpolation:

1. When computing `denoms`, the index for comparison was incorrect. It should
    have obtained the index using `chunk_offset * chunk_size`.
2. When computing `final_coeffs`, it attempted to access `chunk` with
    the index of `numerators`, not the index of `chunk`.

In addition to these bugs, the comments provided incorrect descriptions.

Upon examining the computation of numerator parts, it becomes apparent that
this is a slightly modified version of
`UnivariateDenseCoefficients::FromRoots()`. Therefore, we have chosen to use
this code for both better maintenance and bug fixes.