perf(zk): optimize `DivideByVanishingPolyInPlace()`
In #221, we enhanced `GetSuccessivePowers()` to accept
an additional parameter for offset. Leveraging this improvement,
we eliminate redundant `MultiScalarMul()` calls and the allocation
of `powers_of_coset_gen`. Additionally, the `BatchInverseInPlace()` function
is executed in parallel, incurring slight overhead due to
additional allocations. To address this, we move the operation
within the existing `Parallelize()` function and replace it with
`BatchInverseInPlaceSerial()`, which excludes these allocations.
Lastly, since `base::Parallelize()` also involves allocations of
the vector of `absl::Span`, we opt for `OPENMP_PARALLEL_FOR` instead.