fix(zk): read the commitments to the advice columns in the correct order
Unlike other columns, the advice columns have their own phases. The phase
iterates from 0, and the advice columns are committed when the current phase
is equal to the phases of the advice columns. Therefore, commitments
are created in the order of phase. However, commitments should be written
to the proof in the order of column. For example, if there are commitments
{C₀, C₁, C₂} corresponding to columns 0, 1, 2 and phases {0, 1, 0},
previously, the proof was read in the order of C₀, C₂, C₁,
even though the correct order should have been C₀, C₁, C₂.
In this commit, the order to read the columns has been corrected.