perf: use `Point` instead of `base::DeepRef<Point>`
Initially, we aimed to save memory by avoiding duplicated points.
This led us to introduce the `PointSet` class to check
for the existence of elements. However, the duration for the opening
is not significant, and the size is not substantial,
so optimizing for memory is not worthwhile. Instead, the focus
should be on reducing the time when reviewing the benchmark results.