perf: adopt `c = a op b` style over `c = a; c inplace-op b`
This commit enhances performance by utilizing the `c = a op b` style instead of
`c = a; c inplace-op b`. While the former style is nearly as performant as
the latter when `a` is movable, it becomes inefficient when operating on
prime fields or points on elliptic curves that are not able to be moved.
This change ensures better performance and readability, reducing 2 lines to 1.
Note that irrelevant changes may occur due to this adjustment.