chore(math): update comments to contain montgomery form & fix minor errors
For the 31-bit prime field, the montgomery form is `a * R`, where R is 2³²,
so value `a` can be in the range {0, ..., p}, where p is modulus.
The comments are also updated to handle this case.

Additionally, this commit also contains minor fixes, such as changing `P`
to `p` and removing commas and periods.