perf(zk): optimize `CreateGrandSumPoly()`
This optimization reduces the memory allocation for `CreateGrandSumPoly()`.
Previously, it allocated O(3 * n + usable_rows) memory for every call.
(Note that if the openmp is enabled, there is an additional memory allocation.)
Now, the memory requirement is reduced to O(2 * usable_rows) and is allocated
only once and shared across calls, improving performance and efficiency.