feat!: treat prime field with hex string directly when (de)serializing json
BREAKING CHANGE: This update simplifies the (de)serialization of prime fields
in JSON. Previously, a prime field was (de)serialized as an object
with a member 'value'. Now, it is (de)serialized directly as a hex string.
For example, the value 3 was previously serialized as `{"value": "0x3"}` and
is now serialized as "0x3".