refac(math): optimize the use of `swapElements()`
The size of `evals.evaluations_` is less than or equal to `this->size_`. In
the previous logic, if the size of `evals.evaluations_` is small,
the remainder is filled with `F::Zero()`, and there is no need to perform
swap operations on the part filled with `F::Zero()`. If we apply this
commit, we can avoid unnecessary swaps on parts filled with `F::Zero()`.