feat(circom): repeat only prove part
Updated the proving process to repeat only the prove part,
excluding the zkey parsing. This change is made because
the zkey parsing takes significant time, but typically
this is not the part users want to measure. Additionally,
as rapidsnark does not include zkey parsing,
this adjustment ensures fairness in benchmarking.

Related: #460