perf(halo2): apply batch commitment in `create_proof()`
Previously, `create_proof()` committed to polynomials or evaluations
and then performed batch normalization. This approach was inefficient
on the CPU (though acceptable on the GPU, as `ProjectivePoint` was returned).
The inefficiency arose because `commit()` returned `PointXYZZ`,
which then needed conversion to `ProjectivePoint` and finally to `AffinePoint`.
This process has now been optimized by directly converting `PointXYZZ` to
`AffinePoint`, thus improving performance.