feat: update `CosetLDEBatch()` to accept const and move references
The input matrix is modified after running `CosetLDEBatch()`,
rendering it unusable. Therefore, it makes more sense for the function to
accept a move reference. However, for benchmarking, the input matrix needs
to remain unchanged. Previously, the input matrix was copied
at the caller site. With this update, the function will automatically
handle copying based on the input type.